implementation finished (2)

+++ db/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Date, Interval, ForeignKey, JSON
+from sqlalchemy import create_engine, MetaData, Table, Column, BigInteger, Integer, String, Date, Interval, ForeignKey, JSON
 from sqlalchemy.orm import relationship
 from sqlalchemy.ext.declarative import declarative_base
 
@@ -7,8 +7,8 @@ Base = declarative_base()
 
 class User(Base):
     __tablename__ = 'user'
-    id = Column(Integer, primary_key=True)
-    state_id = Column(Integer, ForeignKey('state.id'), default=None)
+    id = Column(BigInteger, primary_key=True)
+    state_id = Column(BigInteger, ForeignKey('state.id'), default=None)
     session_data = Column(JSON, nullable=False)
 
     def __repr__(self) -> str:
@@ -17,7 +17,7 @@ class User(Base):
 
 class State(Base):
     __tablename__ = 'state'
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     name = Column(String, unique=True, nullable=False)
 
     def __repr__(self) -> str:
@@ -26,8 +26,8 @@ class State(Base):
 
 class Medication(Base):
     __tablename__ = 'medication'
-    id = Column(Integer, primary_key=True)
-    user_id = Column(Integer, ForeignKey('user.id'), nullable=False)
+    id = Column(BigInteger, primary_key=True)
+    user_id = Column(BigInteger, ForeignKey('user.id'), nullable=False)
     name = Column(String, unique=True, nullable=False)
 
     def __repr__(self) -> str:
@@ -36,13 +36,13 @@ class Medication(Base):
 
 class Prescription(Base):
     __tablename__ = 'prescription'
-    id = Column(Integer, primary_key=True)
-    user_id = Column(Integer, ForeignKey('user.id'), nullable=False)
-    medication_id = Column(Integer, ForeignKey('medication.id'), nullable=False)
+    id = Column(BigInteger, primary_key=True)
+    user_id = Column(BigInteger, ForeignKey('user.id'), nullable=False)
+    medication_id = Column(BigInteger, ForeignKey('medication.id'), nullable=False)
     start_date = Column(Date)
     end_date = Column(Date)
     dose = Column(Integer, nullable=False)
-    event_id = Column(Integer, ForeignKey('event.id'), nullable=False)
+    event_id = Column(BigInteger, ForeignKey('event.id'), nullable=False)
     time_delta = Column(Interval)
 
     def __repr__(self) -> str:
@@ -53,7 +53,7 @@ class Prescription(Base):
 
 class SpecialCondition(Base):
     __tablename__ = 'special_condition'
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     name = Column(String, unique=True, nullable=False)
 
     def __repr__(self) -> str:
@@ -62,8 +62,8 @@ class SpecialCondition(Base):
 
 class PrescriptionConditions(Base):
     __tablename__ = 'prescription_conditions'
-    prescription_id = Column(Integer, ForeignKey('prescription.id'), primary_key=True, nullable=False)
-    condition_id = Column(Integer, ForeignKey('special_condition.id'), primary_key=True, nullable=False)
+    prescription_id = Column(BigInteger, ForeignKey('prescription.id'), primary_key=True, nullable=False)
+    condition_id = Column(BigInteger, ForeignKey('special_condition.id'), primary_key=True, nullable=False)
 
     def __repr__(self) -> str:
         return f"<PrescriptionConditions(prescription_id={self.prescription_id}, condition_id={self.condition_id})>"
@@ -71,7 +71,7 @@ class PrescriptionConditions(Base):
 
 class Event(Base):
     __tablename__ = 'event'
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     name = Column(String, unique=True, nullable=False)
 
     def __repr__(self) -> str:
